Better stories: more places, plural shapes, a crisis every time — and green CI on every platform

Finishes sub-project 3 of the combined movie maker. Measured over 200
default-length films against main: two locations became three to five (3.67
average), one beat order became three, and films reaching a crisis went from
0/200 to 200/200 — which is what finally puts the handheld camera, the Dutch
tilt, the broken poses and the full-band score into an ordinary film rather
than only the rare long one. Exterior scenes also stopped using interior
wording (238 occurrences per 400 films down to 6).

Also repairs the engine's cross-platform CI, which had been red on every push
and was hiding real defects. Windows, macOS and the clang-tidy gate are green
for the first time; the unit suite had never actually executed on Windows or
macOS because of a wrong ctest flag, and now runs 930,169 checks there.

Two genuine bugs surfaced from that work: LinearArena refused every 16-byte
aligned allocation under MSVC, and 91 files used std::min/max without
including <algorithm>.

+++ apps/behavior/main.cpp
@@ -110,7 +110,9 @@ int main(int argc, char** argv) {
     for (int i = 0; i < kAgents; ++i) {
         auto ap = std::make_unique<Agent>();
         Agent* a = ap.get();
-        a->wp = static_cast<uint32_t>(i % patrol.size());
+        // i is int, patrol.size() is size_t: widen before the modulo rather than letting it
+        // convert. i is 0..count-1, so the value is unchanged.
+        a->wp = static_cast<uint32_t>(static_cast<size_t>(i) % patrol.size());
         a->body.pos = patrol[a->wp];
         a->body.maxSpeed = 5.5f + static_cast<float>(i % 2);
         a->body.maxForce = 28.0f;
